Add Addition class and corresponding unit tests
Implemented the Addition class, which is responsible for providing binary addition functionality in the Forth interpreter. This class extends ForthBinaryOperator and overrides the accept method to perform addition. Unit tests are also added to ensure the correctness of the implementation. This change allows the interpreter to effectively handle addition operations in Forth language programs.