Add ForthBinaryOperator and ForthWord classes, update pom.xml
Introduced two new classes, ForthBinaryOperator and ForthWord, in the exercism-forth project. These classes are designed for integrating the Forth language's word functionality and implementing binary operators respectively. Additionally, added dependencies (JUnit and Mockito) in the pom.xml file for testing and modified the build plugins to include the Jacoco code coverage tool. These changes are done to facilitate testing and ensure code quality as we add more features to the Forth interpreter.