Add maven-failsafe-plugin and update jacoco-maven-plugin in pom.xml
Added the maven-failsafe-plugin that is designed to run integration tests to verify overall system stability and ensure interoperability between components within the system. The version of jacoco-maven-plugin, a code coverage library, was also updated from 0.7.7.201606060606 to 0.8.11 for ensuring accurate testing and improving test quality. These changes are made to improve software quality and reliability.