Add pom.xml for the legacy-forth project
Added the Maven pom.xml file to manage the legacy-forth project. This XML file is the central configuration file for the Maven build system. It contains settings like groupId, artifactId, and version which are required to uniquely identify the project. Also, it defines the version of Java(17) and the character encoding(UTF-8) to be used in this project. This addition will help in the seamless building of the project.