Refactor Maven configurations for enhanced report generation
The pom.xml configurations were refactored to improve the Maven build process and the generation of test reports. Specifically, the configurations for the 'jacoco-maven-plugin' were restructured, and extraneous elements were removed to streamline the process. Additionally, a 'sonar.language' property was introduced to ensure the correct identification of the project language as Java. Lastly, two Maven plugins were removed from the global scope and instead placed within a profile named 'coverage'. This profile is dedicated to generating a test coverage report using the 'jacoco-maven-plugin'. These changes aim to enhance the quality and comprehension of generated reports.