Remove Mockito dependencies, add XML reporting to pom.xml
The Mockito dependencies were removed from the build files as part of a larger refactoring. It was deemed that these dependencies were no longer necessary for our testing strategy for the ForthEngine. As part of the exact same refactoring, a change was also made to the pom.xml file to include XML reporting. This will allow to capture the test results in a structured format that can be easily processed by other tools.