Update Jacoco report path and remove redundant property
The change modifies the Jacoco report path in the project properties of pom.xml file to refer to an aggregated report in a different module. This was updated to ensure that code coverage metrics include all project modules in their calculations. Additionally, a duplicate Sonar report property was removed as it was redundant and potentially confusing.