Add ForthEvaluator class and unit tests; Update pom.xml
Added a new class ForthEvaluator to perform command evaluations. This class supports a certain set of commands like addition, subtraction, multiplication, etc. Also implemented related unit tests to ensure the correct working of the evaluator. The pom.xml file was modified to include the 'forth-legacy-one' module which contains the added ForthEvaluator class and its unit tests to the list of maven modules in the project. This was essential to ensure the new class and associated unit tests are part of the build lifecycle, hence get compiled and tested during a maven build.